add theme for JOSS
JOSS (The Journal of Open Source Software) uses a TeX template [1] to
compile submissions into PDFs.

Add a MakiePublication theme with figure size based on the output of
```
linewidth=\uselengthunit{pt}\printlength{\linewidth}
\\
linewidth=\uselengthunit{in}\printlength{\linewidth}
```
within a `figure` environment in this template.

The output of this is
```
linewidth=5.36066 in
linewidth=387.33861pt
```

[1]: https://github.com/openjournals/inara/blob/dc6adfd5b65c304154131f3f4fe08aba837677f7/data/templates/default.latex